fix: Trade/Position Integer 컬럼을 Float으로 변경, 무한루프 방지
- Trade.entry_price, entry_qty, exit_price, stop_price, profit_loss: Integer → Float
  US 소수점 가격($180.50)과 소수점 수량(1.5주) 저장 시 정수 잘림 버그 수정
  KR 기존 데이터는 정수값이 Float으로 저장되므로 하위 호환성 유지
- Position.qty: Integer → Float (US 소수점 수량)
- overseas_market_data.get_daily_ohlcv: while 무한루프 → for 유한루프

+++ repository/models.py
@@ -48,13 +48,13 @@ class Trade(Base):
     stock_name = Column(String(50))
     status = Column(String(10), default=TradeStatus.OPEN, nullable=False)
     entry_at = Column(DateTime, nullable=False)
-    entry_price = Column(Integer, nullable=False)
-    entry_qty = Column(Integer, nullable=False)
+    entry_price = Column(Float, nullable=False)   # KR: 원 정수, US: USD 소수점
+    entry_qty = Column(Float, nullable=False)     # KR: 정수, US: 소수점 수량
     exit_at = Column(DateTime)
-    exit_price = Column(Integer)
-    stop_price = Column(Integer)
+    exit_price = Column(Float)
+    stop_price = Column(Float)
     stop_type = Column(String(20))   # fixed | trailing | atr
-    profit_loss = Column(Integer)
+    profit_loss = Column(Float)
     profit_loss_pct = Column(Float)
     exit_reason = Column(String(50))  # stop_loss | take_profit | signal | emergency
     signal_id = Column(Integer)
@@ -68,7 +68,7 @@ class Position(Base):
     market = Column(String(2), default="KR", nullable=False)  # KR | US
     stock_code = Column(String(10), nullable=False, unique=True)
     stock_name = Column(String(50))
-    qty = Column(Integer, nullable=False)
+    qty = Column(Float, nullable=False)      # KR: 정수, US: 소수점 수량
     avg_price = Column(Float, nullable=False)
     stop_price = Column(Float, nullable=False)
     stop_type = Column(String(20), default="fixed")
